bookbook/BACKEND/requirements.txt
Django>=5.0,<5.3
djangorestframework>=3.15
asgiref>=3.9
sqlparse>=0.5
django-cors-headers>=4.9
typing_extensions>=4.14
tzdata>=2025.2